+++ tests/test_routing.py
@@ -104,12 +104,13 @@ class TestRoute(TestCase):
         self.handler = f
 
     def test_ok(self):
-        route = Route(path="/user/bob", fn=self.handler, opts={"foo": "bar"})
+        route = Route(path="/user/bob", fn=self.handler, parents=["/user/{name}"], opts={"foo": "bar"})
         self.assertEqual("/user/bob", route.path)
         self.assertEqual({}, route.convertors)
         self.assertEqual({"foo": "bar"}, route.opts)
         self.assertEqual((True, {}), route.match("/user/bob"))
         self.assertEqual((False, {}), route.match("/user"))
+        self.assertEqual(["/user/{name}"], route.parents)
 
     def test_convertor_str(self):
         route = Route(path="/user/{name}", fn=self.handler)